Bootstrap environment for virtual deployment
If the deployment is not done on baremetal, we need to install
libvirt, its dependencies and ensure some of the adjustments are
done beforehand so the subsequent tasks do not have to deal with
issues.

Please note that this change is only applicable for bifrost.
If new tools are introduced in future, the playbook and the role
bootstrap-virtual-environment need to be made common.

Change-Id: Ifa8ebb08ed89f627814c46e84a822a3fda1fcb39
Signed-off-by: Fatih Degirmenci <fdegir@gmail.com>