Add PDF, IDF, and SDF
This change adds PDF and IDF for virtual deployment and a placeholder
for SDF.

- PDF: POD Descriptor File. This file describes to capabilities of the
nodes within virtual POD. It is based on OPNFV PDF. Baremetal PODs will
need to have similar PDF files as well so we can programmatically
do stuff with them.
- IDF: Installer Descriptor File. This file is a glue between PDF and
OpenStack and Kubernetes installers to manipulate, configure, and
operate on nodes and networks. It is based on OPNFV IDF.
- SDF: Scenario Descriptor File. This file is currently empty and it
will contain lsit of scenarios, their locations, versions and so on.

Change-Id: If1f20ac63c2fe86ca0a23b34df4e34d15cc65049
Signed-off-by: Fatih Degirmenci <fdegir@gmail.com>