Add the impi privilege role to inventory file
This change has added a 'priv_level' variable into idf file which
can be used in inventory creation step to define the ipmi privilege
level. If there is no priv_level variable in idf, the ipmi_priv_level
will be set to default value(ADMIN).

The reason why we make this change is, in RANi lab, provision will
fail if we don't set ipmi_priv_level to OPERATOR.

Change-Id: I1b4d3e41923ba993d4294d045e9edd1261f93178