Separate admin and public networks and interfaces
This change literally rollsback the earlier change that removed
admin network and combined admin and public networks. [1]

The reason for rolling it back is that in certain cases, the network
that is used for PXE booting might not have internet access, making
it necessary to use another network for internet access.

Also, this change makes it more consistent with the terminology;
admin network is used for PXE booting which is what we will be doing
within bifrost inventory and during the provisioning.

[1] https://gerrit.nordix.org/#/c/405/

Change-Id: I5d62aaa216ff8ef6652c6a87d3c9e7139e03c92a