Configure Kubespray installer to install Kubernetes
This change configures everything to install Kubernetes using
Kubespray.

As noted in the corresponding files, there are things that could
be pushed into the scenario and it will be done once the basic
installation works from a single place.

Apart from that,. there is a workaround in place to avoid the
Kubespray bug.

https://github.com/kubernetes-sigs/kubespray/issues/4300

Change-Id: I5e3a6831bd371f49b57c41f106e3b854f835f911