Introduce Ubuntu1804 support and set it as default
It seems we don't need to move to netplan as the interface
configuration files and the commands to bring down/up interfaces
work without issues on Ubuntu1804 as well.

This change just symlinks ubuntu16 related task files and templates
for ubuntu18. The reason for this is that we might still need to
adapt netplan so keeping the structure as is might be useful for
future.

Apart from that, there is yet another workaround for DNS for /etc/resolv.conf
as the content of it seems to be generated randomly for some reason. This
needs to be investigated further.

Docker official GPG key needs to be added as well.

Change-Id: Id8dfaa3dbed4c2675bc839040b573ba2aab226c5