Add support of dhcp and manual interfaces in configure-targethosts role
Field "address" in PDF is reused.
If it is set to "dhcp", the interface is set to "dhcp" enabled.
If it is set to "manual", the interface will be disabled by setting to "manual".

Also add a small fix the namespace in network interface file
(it is probably not taking effect anyhow).

Note:
There's no example of "dhcp" network in pdf/idf, we have issues
in dhcp handling in libvirt network as dhcp service was killed
during provisioning period.

Change-Id: Ibb542f6f45c6e1b24f23a506dc885ef8319ce40c