[INF-63] Fix dns configuration
In the OS image, systemd-resolved is used as local dns manager,
therefore in this commit we config the DNS via the
corresponding configuration file: /etc/systemd/resolved.conf

Removed the changes to network interfaces file as this dns manager
will not respect that. Removed also to the changes to resolve.conf
as the file will be managed by the dns manager.

At the same time, restore the idf back to support multiple dns
servers as a list.

Also, to make the host server be less resource demanding,
modified the pdf file with reduced hard disk and memory.

Change-Id: I0ce7beb644bb5e2757bb62da46aca6d3f4e86e1f