Install ironic components using pip
Installing ironic and the other ironic components from source is
something that comes from OPNFV XCI as we provided feedback to
OpenStack community by testing stuff from source.

In our case, we are not providing feedback so using pip should be
sufficient. This obviously makes us a bit vulnerable since there
is no way to pin ironic stuff in bifrost but at the same time, this
is not the only situation we will be open to this type of issues.

Apart from that, there is no way to install ironic using pip so
we still install ironic from source by pinning it so this should
provide some kind of protection.

Finally, we also stop using upper-constraints as it is not necessary
in our context and let bifrost itself handle it.

Change-Id: Ibc89c67d684379ad99cd19de55431848e97cc9ab