Use prebuilt images for provisioning
This change makes it possible to use prebuilt images during
provisioning with bifrost. Here are the details:

- a new var, use_prebuilt_deployment_image, is introduced to
control the use of prebuilt deployment image. It is set to
false by default to ensure when the engine is run manually,
deployment image gets built with the user's ssh public key.
- The var use_prebuilt_deployment_image can be overriden by
setting the env var USE_PREBUILT_DEPLOYMENT_IMAGE if it is
run within CI so the build of the deployment image can be
skipped.
- Since the deployment image contains ssh public key, if
USE_PREBUILT_DEPLOYMENT_IMAGE is set to true on a different
CI system, arm_url needs to be overriden to get the right
image for the systems other than Nordix one.
- IPA images are downloaded from artifactory by default and
the var use_prebuilt_deployment_image doesn't take effect.
- The folder /httpboot is not touched except removal of
the deployment and IPA images appropriately so if one doesn't
want to use prebuilt deployment image, the existing one will
be left as is. If it doesn't exist, it gets built. IPA images
are always removed and redownloaded.

Change-Id: Ice16bc3ec0fbb134573ed696d7d8e5ec6cd86d49