bugfix: Correct the network configuration order
Today we remove all the old configuration files,
write the new configuration files and bring the interface
down and up. In this way dhclient process will be left over
for the interfaces that are changed from dhcp to static or manual,
as the ifdown command does not know the interface was configured
as dhcp.

Correct order of the configuration is to bring down the interface,
update the configuration file and then bring it up.

The issue in dhcp client is suspected to be connected with the dns
instability issue in some lab as the dhcp client is keeping receiving
dns information which caused the dns manager restarts all the time.

Change-Id: I78ba7b8e9d0ac3785d87a8cd794a77b94a599755