Add post-deployment playbook and role
The post-deployment playbook and associated role installs kubectl
using package manager and copies kubernetes config fine in ~/.kube
in order to be able to interact with kubernetes cluster.

The post-deployment playbook and the role can be used for other and
perhaps scenario specific tasks as well with a corresponding post-deployment
playbook in scenario directory if it exists.

Change-Id: Ifcf9925d126141e34114a80b5e168c8a8b986234