Add root device selection during provisioning
This solves the issue that the OS is not always installed
in the primary disks.

When the machine boots up the order of the disks is random,
so ironic needs this property (root_device hint) to know
exactly in which drive the system should be written to.

Otherwise, there can be problems in baremetal when
the disk or its controller does not support handling an OS.

This is not yet supported by the latest release of Ansible,
but will be included in the next one v2.8.
Then this workaround should be substituted by adding
this same property with Bifrost inventory generator.

Change-Id: I646613f1f839cdc08c673bc5e70d1d1ab8884983