Remove gateway and dns configuration in provisioning
During provisioning, we do not access to external network,
therefore we do not need to configure the gateway and dns.

Also the provisioning process will configure these two
parts in the PXE interface, which is not always correct.

In this patchset, this part is removed. Should be no impact
as we configure our network separately in later stages.

In the same script, also adjusted the order of the root device
selection so we always print out after adjustment.

Change-Id: I65dfbdbbddd2668de8b1fecf58e270424f20bb94