Update to configure the PXE network with dhcp in glean
Double checked, gateway will anyhow be provided to the network configuration,
since both bifrost and glean will write the gateway without checking if it is
empty. This still causes instabilities if the network is not the gateway network.

Update to use dhcp instead for PXE network, as it is still available
at the time being before we cleanup the dns server after completion.
And then we will configure network correctly in seperate steps.

Change-Id: I25dd45da6078d24e31a45b4a680d601147f17e3a