Take infra/swconfig repo into use
This change makes it possible to decouple scenarios themselves
from the engine by adding mechanism to bring the scenario specific
roles from infra/swconfig repo.

The way the mechanism works is
- a scenario is created in infra/swconfig repo, following the directory
structure scenarios/<scenario name>/<installer type>/role/<scenario name>
- the created scenario is declared in engine/engine/var/sdf.yml
- engine copies sdf.yml into engine/.cache/config/sdf.yml
- engine verifies the scenario to be deployed to see if it is supported
by the selected installer and on the distro
- engine clones the scenario from the source repository into engine/.cache/repos,
checks out specified version
- engine bootstrap-scenario.yml playbook uses the scenario role so
scenario specific stuff can be done before the execution of actual install
command

One can use refspec to deploy a scenario that is still open on Gerrit like how
it is done in this change for the scenario k8-calico-nofeature. Once the
corresponding change gets merged, the version to use for the scenario will be
switched to master.

A more detailed guide, listing step by step instructions for scenario development
and testing either locally or on Gerrit will be made available once the implementation
is done..

Change-Id: I8d080ec3891fe27ed2cf595e5d8f81b12fa090d8