Bump kubespray in order to bring k8s v1.14.0
This change bumps kubespray sha to 05dc2b3 in order to bring k8s
v1.14.0.

Previous attempt failed due to the task that makes k8s admin.conf
available on jumphost. It looks like k8s v1.14.0 kubeadm behavior
changed, resulting in inclusion of erroneous stdout line in the
file, causing failures when kubectl is attempted to be used or
during testing with functest and yardstick. [1]

    [kubeconfig] Writing "admin.conf" kubeconfig file
    apiVersion: v1
    clusters:
    - cluster:
    ...

In order to accommodate this error, a task in engine post-deployment
role is created as a workaround.

[1] https://github.com/kubernetes-sigs/kubespray/blob/master/roles/kubernetes/client/tasks/main.yml#L78

Change-Id: I1238806983ecb763037f05d38bf2af2b4ae698f4