Enable setting scenario specific variables using overrides
Scenarios might require the use of specific versions of the
components that are different from the ones that are pinned
in engine repository. In order to make this possible, the
variables can be overriden by the scenario itself before
running the rest. An example to this is ONAP based scenarios
which require different versions of k8s, kubectl, helm, and
docker.

This change enables this by sourcing the file named overrides
before running anything else so the versions pinned in
var/versions.yml are set using the ones coming from the scenario.

As having scenario overrides is optional, this file does not
have to exist.

overrides file  must be located as below in order for this
to work
  <swconfig_repo>/scenarios/<deploy_scenario>/<installer_type>/overrides
This file then becomes available in engine cache as below while
bootstrapping swconfig repository
  <engine_cache>/repos/swconfig/scenarios/<deploy_scenario>/<installer_type>/overrides

Change-Id: I405a9f1ab39618d08a98e788ac4cafbc99427f84