Enable running scenario specific post deployment tasks
Scenarios might have additional tasks that need to be run
after the completion of the base scenario installation. An
example to this is ONAP based scenarios which require k8s
to be installed first and then ONAP using OOM.

This change enables this by running the tasks from post-deployment.yml
that may be located in scenario role folder. As having post deployment
tasks is optional, this file does not have to exist.

post-deployment.yml playbook must be located as below in order for this
to work
  <swconfig_repo>/scenarios/<deploy_scenario>/<installer_type>/role/<deploy_scenario>/post-deployment.yml
This playbook then becomes available in engine cache as below while
bootstrapping swconfig repository
  <engine_cache>/repos/swconfig/scenarios/<deploy_scenario>/<installer_type>/role/<deploy_scenario>/post-deployment.yml

Change-Id: Iebc862ec27ec742547bc80e8a9e718f39bd12ee2