Refactor heat template and environment file
List of changes:
- name instances as <instance_name><index>.<stack_name>
examples are
    master0.example-stack
    worker0.example-stack
- name volumes as volume<instance_name><index>.<stack_name>
    volume.master0-example-stack
    volume.worker0-example-stack
- create new keypair specific to the stack and use that one instead
of an existing keypair. this is done by creating the keypair using
os_keypair ansible module together with user's existing public key
on filesystem. the reason for this is that heat get_file function
does not support the tilde or use of environment variables within
the template to get file contents while specifying the running user's
public key location. keypair is named as keypair-<stack_name>.
  keypair-example-stack
- create network, subnet, router and name them network.<stack_name>,
subnet.<stack_name> and router.<stack_name> respectively. examples are
subnet.<stack_name>. examples are
  network.example-stack
  subnet.example-stack
  router.example-stack
- move cloud sepcific parameters to heat environment file, heat-environment.yaml

Change-Id: Iaa6420d1ce7ad619d54f84707186515c0e02ed5a