Add jumphost to heat based deployment
As the heat based deployments use OpenStack clouds,
it is important to have a jumphost between the user
and the target deployment so access, deployment, operations
and so on could be done on jumphost.

Another importance of having jumphost is to ensure the
target deployment is put in correct security groups with
as few as possible open ports to prevent unauthorized access
to the deployment. The only full access to the target deployment
will then be available from jumphost which again has as few
ports open as possible.

Change-Id: I6eaf78e217b4dbb671201686f4c3a755afbfab6d