Bump bifrost and ironic shas
Bifrost introduced support for Ubuntu 1804 so this change bumps
bifrost and ironic versions to bring it in.

There seems to be a problem with wait_for module on ubuntu 18.04
as the task in bifrost-enroll-deploy.yml times out when used with
hostnames so we switch to ip address taken from bifrost baremetal.yml
file with ansible_ssh_host.

Change-Id: I5fb78682ec725e2d0cc9bf47f01066ddbf9a70b0