Rename main.yml playbooks to pre-deployment.yml
Scenarios currently have main.yml playbooks that contain
tasks to run pre-deployment of the platform. This change makes
it possible to rename main.yml playbooks in scenario roles
to pre-deployment.yml so it becomes more visible that the
tasks in pre-deployment.yml playbook are run before the actual
deployment.

A related change is sent to all the scenarios to rename
main.yml playbooks to pre-deployment.yml.
  https://gerrit.nordix.org/#/c/1660/

Change-Id: Ie9f984e67ce8f6f0dc5c8a60e0f3984cb84cca75