Move remote user to generated kubespray inventory
We operate on target nodes as root user so this change sets
ansible_user to root in generated kubespray inventory and removes
remote_user: root from the corresponding playbooks to ensure
we have this configured in single place.

As a side benefit and since we operate on target nodes as root user
anyways, creation of a new user on target nodes is also removed as
we don't need it anymore.

Change-Id: I852ba6042e8e317c55d6598133d28f0fb8d6a29f