Refactor pre-deployment to run on all hosts
Current implementation of the pre-deployment mechanism does
not the possibility of running pre-deployment tasks on any
hosts within inventory. However, it is possible that we need
to run pre-deployment tasks for more than just localhost for
certain scenarios such as onap-oom-nofeature.

This change switches hosts: localhost to hosts: all in pre-deployment.yml
file. All the existing scenarios must be adapted to this way of
running pre-deployment tasks by choosing the right host per playbook
or task in them.

See the change adjusting existing scenarios.
https://gerrit.nordix.org/#/c/1767/

Change-Id: I44ac5d318333a978c23b7f1e1ea9e44bf822374f