Refactor post-deployment tasks and role
Rook related tasks and the tasks to configure localhost
for kubectl and helm are spilled over under different
playbook and role. This makes it difficult to update
them especially during the introduction of heat.

This change refactors the tasks and the role to make
it easier to develop things further and ease the effort
to introduce new tasks into common post-deployment.

Change-Id: I325ae7ee438ef355b64e5d4be92349e348c542e6