Create security groups and place instances in
When the deployment is done on the instances created using Heat,
it is important to lock down instances to external access.

This change creates 2 security groups, internal_security_group and
external_security_group, with the details below.

- internal_security_group: all the instances including jumphost and
the instances where the stack is deployed are placed in this group. all
ingress ports are open for tcp & ucp between the instances located in
this security group.
- external_security_group: only the jumphost instance is located
in this group. only the tcp 22 (ssh) ingress port is open so users
can ssh to this instance to access to the target deployment.

internet <-- external_security_group (port: 22)  --> jumphost <-- internal_security_group (ports: all) --> SUT

Change-Id: I6b825c4a9b55481c197c617a850e0f7b5c2eeb8d