Flush cache for kubespray
We are experiencing issues with kubespray which we can't explain
why. The problem occurs when engine is executed on a machine that
was used for executing engine previously and kubespray complains
that node ipv4 addresses don't match.

This is an attempt to see if kubespray is leaving cache behind
which we don't know thus don't clean.

As CI always uses fresh machines, the patch will probably pass
CI so this patch needs to be verified manually to see if the new
option would help with the problem.

Change-Id: I2fe10dd8ebd985c3f789edfe63e155f6e4451f5f