Move app installation playbook into engine
Generate the app installation playbook while running
the bootstrap-swconfig role.

The reason to change this is because ansible import_playbook
does not support loops, and we need to check each and every
app if it needs to be installed and import them.
For e.g.
- import_playbook: "istio/{{ installer_type }}/playbooks/install.yml"
  when: "'istio' in apps"

Also, if the playbook itself does not exist for all the installers
this will not work. For e.g. istio is not supported by kolla, so
all scenarios using kolla installer will not work with this approach.

Change-Id: I950f089f1398ea55e15a9c4914f117ee8e854e90