Initial implementation of offline engine deployment
A new option has been added "-x" to execute this use case.
It will use a predefined tar.gz package that contains all
the necessary files to perform the installation of
k8s-calico-nofeature scenario and has to be downloaded in
advance. It has been documented the structure of this file
and how to fetch the dependencies in an online machine.

Apt dependecies are handled by apt-cacher-ng that sets up
a proxy in localhost. This is used by the newly created nodes.
Kubespray cache (binaries and docker images) is copied to the
nodes and the playbook is configured to use those files.

In post deployment, the remaining of Docker images are pre-
loaded in the target machines. A better solution would be
to establish a local docker repository, but it is left out
of this patchset.

Change-Id: I3155e3bca56435a97f165a266935e7fec49f1a4d