Move python and python-dev installations out of heat templates
Due to stack creation issues on public cloud, installation of python
and python-dev has been moved out of heat templates. The reason for this is
that during initial boot of the instances, Ansible prerequisites python and
python-dev are installed using boot script. Due to the reasons we can not
explain and perhaps because of network issues on public cloud, apt fails,
complaining about corruption - checksum mismatch. In turn, problematic
instance(s) can not send completion signal at the end of boot phase, resulting
in timeouts in stack creation thus complete failure of deployments.

This change moves python installation into a separate script which can
then be executed by Ansible on the instances since script module does not
require Python on the remote system, much like the raw module. Once
python and python-dev are installed on the instances, we can continue
with the rest as usual.

Change-Id: I1f2d96d84633b5b5f6fc0c3df02b033c40891249