Use tox for running static code analysis
Engine has already been refactored to fix ansible-lint and
shellcheck issues. However, running static code analysis is
not straightforward as it requires installation of dependencies,
finding what to analyse, and using the right combination of arguments
for the utilities.

This change introduces tox for automating static code analysis both
for developers and for CI. In addition to tox itself, requirements
are collected in test-requirements.txt file to manage them. Documentation
is also created to describe how one can run tests locally. CI will
be adapted to use tox once this change goes in.

Change-Id: I75fe8f22d3854e85aa0023465454a9125c1964b8