group_vars: Refactor bootstrap-configuration
This change refactors engine bootstrap-configuration
playbooks to use group_vars instead of using vars
from engine/engine/var directory.

As bootstrap-configuration is run on localhost, an
inventory file, engine/engine/inventory/localhost.ini,
is created to enable the use of group_vars, corresponding
ansible-playbook commands updated to use it, and include
directives in playbooks are removed.

This is the first change in series of changes to take
group_vars in use. The transition to group_vars will
be done in a stepwise approach to avoid one big change
that will potentially make the troubleshooting and
code review difficult. Apart from that, the current vars
files and the mechanism in untouched parts in individual
patches are kept as is to keep backward compatibility
till the end. Once all parts are adapted, old mechanism
will completely be removed in the final change of the
series.

Previously, the vars needed for engine were kept in
engine/engine/var directory which required us to include
these files in all the playbooks. Even though this worked
ok, it was unnecessary complication as Ansible group_vars
provides much cleaner and standard way to access the vars.
This will also make it easier for us to access vars from
scenario playbooks without dealing with extra steps to
include them.

Change-Id: I5cae36d17f6e5203dc2d1f432c7e0ed3b00d40c0