group_vars: Refactor provisioner/heat
This change refactors heat provisioner playbooks and
roles to use group_vars instead of using vars from
engine/engine/var directory.

As heat/provision.sh is run on localhost initially,
it uses the newly created inventory file for it to
enable the use of group_vars. The full inventory generated
for created stack is also placed in engine/engine/inventory
as inventory.ini instead of putting it under
{{ config_path }}/inventory.ini. Playbooks and roles are also
updated to remove include directives.

See the first change in series for the reasoning of
switching to group_vars which has additional information
as well.

Change-Id: I830f27bfadcbd7795106986b6f117536d135f8ab