group_vars: Refactor provisioner/bifrost
This change refactors bifrost provisioner playbooks and
roles to use group_vars instead of using vars from
engine/engine/var directory.

As bifrost/provision.sh is run on localhost initially,
it uses the newly created inventory file, localhost.ini,
to enable the use of group_vars. In the last command of
bifrost/provision.sh, bifrost_inventory.py is copied to
engine/engine/inventory to access to group_vars so the
the new full inventory file, inventory.ini, to use during
stack installation can be generated.

See the first change in series for the reasoning of
switching to group_vars which has additional information
as well.

Change-Id: I0cc76c2d2e1fb3b192e395c64f5df476b68a9200