group_vars: Remove code needed for backward compability
This change removes the code that was left to maintain
backward compatibility while implementing group_vars.
With this change, the engine itself totally switched to
group_vars but swconfig/app still relies on engine/var/global.yml
and engine/var/version.yml which will be fixed 2 separate
changes. [0][1]

See the first change in series for the reasoning of
switching to group_vars which has additional information
as well.

[0] https://gerrit.nordix.org/c/infra/swconfig/+/3322
[1] https://gerrit.nordix.org/c/infra/engine/+/3321

Change-Id: I83316a391233b300cbe78ff8912927c1681f57af