group_vars: Remove var files
This is the last change in the series to switch engine to
group_vars entirely. In order for this change to be verified
properly, change sent to infra/swconfig repo must be merged
first. [0]

See the first change in series for the reasoning of
switching to group_vars which has additional information
as well.

[0] https://gerrit.nordix.org/c/infra/swconfig/+/3322

Change-Id: I0f082ff49ce28698eb14d39ecd82df5aa2a7a8a1