Bugfix to use user provided SDF[2]
Force copying the user provided SDF was apparently not enough.
https://gerrit.nordix.org/#/c/infra/engine/+/3331/

SDF information in the user provided sdf.yaml file does not
take effect in this playbook, although we have replaced the
in group_vars/all directory, this playbook is run with whatever
information was read from the default SDF when the playbook was
invoked. So, include the user provided sdf.yaml before running
the rest of the playbook.

Change-Id: Ia226a41f587f5226bd22995a315fc52030bb5faa