Fix yamllint issues and make it blocker
All yamllint issues in yaml files have been fixed except line-length rule
which is excluded. tox.ini is updated in order not to ignore yamllint issues
so if yamllint fails, the change will get verified-1 and will be blocked from
merging. yamllint configuration file, .yamllint, is also created to configure
what is enforced.

As a summary, most frequently occurring issues that are fixed are below.
- truthy value: must be true/false and not yes/no
- indentation: blocks are incorrectly indented mostly
- braces: extra spaces after opening and before closing curly braces are not good
- colons: extra spaces after colons - should be single
- missing document start
- comments: there should be a space after hash

Change-Id: I0d1341f2639585228b1ae8fd6ee073e5cae53604