offline-deployment: Adapt k8s deployment to new packaging
This change updates the functionality introduced by Ignacio
to adapt offline deployment to new packaging and local apt
and docker registry services instead of using apt-cacher-ng
or kubespray local cache. Please note that the change is
not compatible to previous method of offline installation
as it introduces some drastic changes. Also, only bifrost
and kubespray are supported and the work to support heat is
pending. Baremetal is not supported either and will be introduced
once this and other changes this one depends on are merged.

The implementation and usage details are in user-guide.rst
so please refer to it for details.

Change-Id: Ib768476c9bf23b2d772490293bef66e7c055e108