offline-deployment: Set jumphost IP dynamically
For offline deployments, local services are provisioned on jumphost
such as apt mirror, docker registry and so on. In order for target
hosts to use these services, jumphost needs to be added to /etc/hosts
on target nodes.

This change implements fetching and setting jumphost IP dynamically
both for Bifrost and Heat. This will also work for baremetal as well.
For Heat based deployments, Heat templates needed to be updated as
well to get jumphost internal/private IP since the target nodes
can not access to public IP due to security group settings.

Change-Id: I8770ce09554e7eaca998f165f72638c657b7a40e