offline-deployment: Create self extracting installer and documentation
The final archive file is converted to self extracting
installer to make it easier to use it so users can fetch
the file and execute it, getting all the files placed
in right places and informed about what to do next.

Please note that there is a command in build script to
allow testing of the functionality manually while the changes
are reviewed. It will be removed once the reviews are done.

Change-Id: I7effb2e0439380cadcc89a2a8bda901cc1f5ea9a