Pin and use known versions of Bifrost dependencies for online deployments
Currently we do not use pinned versions of Bifrost dependencies for
online deployments and this is problematic since any new commit
coming into the branch bifrost checks out has possibility to break
things for us.

Apart from that, the default checkout behavior of bifrost is to
match the branches of the dependencies it checks out to the bifrost
branch. This is also risky and prevents us from bumping bifrost itself
while keeping the rest on known versions. By using pinned versions for
everything, we can bump versions in a more controlled way.

Change-Id: I40f04d1a2baf3fb5659fafdb51cb661a734a10d7